Configure project as an R package with routine installation; optimize boundary finding and breakpoint file creation; refactor scripts to reduce number of output files; put output files in user-specified output directory; allow key parameters to be passed to function calls rather than hard-coded in source files. Fix (presumed) error in chr coordinate file. Eliminate warnings and errors in R and Python.
