+++ .Rbuildignore
@@ -1,0 +1,4 @@
+^data-raw$
+untracked/*
+^.*\.Rproj$
+^\.Rproj\.user$
