+++ .gitignore
@@ -1,0 +1,4 @@
+.Rproj.user
+.Rhistory
+.DS_Store
+untracked/*
