+++ NAMESPACE
@@ -1,0 +1,29 @@
+# Generated by roxygen2: do not edit by hand
+
+export(do_biscut)
+export(get_chromosome_arms)
+export(get_chromosome_coordinates)
+export(get_gene_locations)
+export(make_breakpoint_files)
+export(plot_BISCUT_results)
+import(data.table)
+import(ggplot2)
+importFrom(grDevices,dev.off)
+importFrom(grDevices,pdf)
+importFrom(graphics,abline)
+importFrom(graphics,lines)
+importFrom(graphics,par)
+importFrom(lifecycle,deprecated)
+importFrom(methods,is)
+importFrom(reticulate,source_python)
+importFrom(stats,ks.test)
+importFrom(stats,p.adjust)
+importFrom(stats,pbeta)
+importFrom(stats,qbeta)
+importFrom(stats,quantile)
+importFrom(stats,runif)
+importFrom(stringr,str_length)
+importFrom(stringr,str_replace)
+importFrom(utils,read.csv)
+importFrom(utils,read.table)
+importFrom(utils,write.table)
