+++ R/BISCUT-package.R
@@ -1,0 +1,27 @@
+#' @keywords internal
+"_PACKAGE"
+
+#' @import data.table
+
+## usethis namespace: start
+#' @importFrom graphics abline
+#' @importFrom graphics lines
+#' @importFrom graphics par
+#' @importFrom grDevices dev.off
+#' @importFrom grDevices pdf
+#' @importFrom lifecycle deprecated
+#' @importFrom methods is
+#' @importFrom reticulate source_python
+#' @importFrom stats ks.test
+#' @importFrom stats p.adjust
+#' @importFrom stats pbeta
+#' @importFrom stats qbeta
+#' @importFrom stats quantile
+#' @importFrom stats runif
+#' @importFrom stringr str_length
+#' @importFrom stringr str_replace
+#' @importFrom utils read.csv
+#' @importFrom utils read.table
+#' @importFrom utils write.table
+## usethis namespace: end
+NULL
