--- docs/SNP6_hg19_chromosome_locs_200605.txt
@@ -1,24 +1,0 @@
-chromosome_info	size	q_start1	q_end1	offset	centromere	cent_mod	p_start	p_end	q_start	q_end	middle_of_arm
-1	249250621	125000000	249250621	0	125000000	0	3218610	120527361	149879545	247813706	124625310.5
-2	243199373	342550621	492449994	249250621	93300000	0	484222	89130753	95341700	242476062	370850307.5
-3	198022430	583449994	690472424	492449994	91000000	0	2212571	90485962	93734671	197538677	591461209
-4	191154276	740872424	881626700	690472424	50400000	0	1053934	48068826	52698757	188763651	786049562
-5	180915260	930026700	1062541960	881626700	48400000	0	914233	45378456	50124856	180360469	972084330
-6	171115067	1123541960	1233657027	1062541960	61000000	0	1014281	57206485	62234533	170903919	1148099494
-7	159138663	1293557027	1392795690	1233657027	59900000	0	705284	57935910	63112265	158385118	1313226359
-8	146364022	1438395690	1539159712	1392795690	45600000	0	617667	42933372	47460945	145232496	1465977701
-9	141213431	1588159712	1680373143	1539159712	49000000	0	789932	38747688	71039044	140938752	1609766428
-10	135534747	1720573143	1815907890	1680373143	40200000	0	415240	39064000	42947970	135225087	1748140517
-11	135006516	1869607890	1950914406	1815907890	53700000	0	456120	50055330	56013798	134142530	1883411148
-12	133851895	1986714406	2084766301	1950914406	35800000	0	889902	33521107	39223566	133161346	2017840354
-13	115169878	2102666301	2199936179	2084766301	17900000	17900000	1	17900000	19450806	114987458	2142351240
-14	107349540	2217536179	2307285719	2199936179	17600000	17600000	1	17600000	20501368	105988038	2253610949
-15	102531392	2409817111	2409817111	2307285719	19000000	19000000	1	19000000	23687685	101884307	2358551415
-16	90354753	2446417111	2500171864	2409817111	36600000	0	653459	31850299	46507665	89379936	2454994488
-17	81195210	2524171864	2581367074	2500171864	24000000	0	987221	21305392	25270517	80917016	2540769469
-18	78077248	2598567074	2659444322	2581367074	17200000	0	329586	14620707	18689747	77109240	2620405698
-19	59128983	2685944322	2718573305	2659444322	26500000	0	284018	24052992	28240494	58878226	2689008814
-20	63025520	2746073305	2781598825	2718573305	27500000	0	455764	26197899	29843973	62219837	2750086065
-21	48129895	2794798825	2829728720	2781598825	13200000	13200000	1	13200000	15347621	47678774	2805663773
-22	51304566	2844428720	2881033286	2829728720	14700000	14700000	1	14700000	17423930	49331012	2855381003
-23							3157107	58317986	62172719	154905589
