+++ man/BISCUT-package.Rd
@@ -1,0 +1,15 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/BISCUT-package.R
+\docType{package}
+\name{BISCUT-package}
+\alias{BISCUT}
+\alias{BISCUT-package}
+\title{BISCUT: What the Package Does (One Line, Title Case)}
+\description{
+What the package does (one paragraph).
+}
+\author{
+\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
+
+}
+\keyword{internal}
