+++ man/filter_big_small.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/helpers.R
+\name{filter_big_small}
+\alias{filter_big_small}
+\title{Filter out largest/smallest tel/cent events}
+\usage{
+filter_big_small(df, telcent_thres = 0.001)
+}
+\description{
+Filter out largest/smallest tel/cent events
+}
