+++ man/get_chromosome_coordinates.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/helpers.R
+\name{get_chromosome_coordinates}
+\alias{get_chromosome_coordinates}
+\title{Get a table of chromosomal p and q arm coordinates}
+\usage{
+get_chromosome_coordinates(genome = "hg19")
+}
+\arguments{
+\item{genome}{Which genome build to use. Currently, only "hg19" is available.}
+}
+\description{
+Get a table of chromosomal p and q arm coordinates
+}
