+++ man/get_gene_locations.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/helpers.R
+\name{get_gene_locations}
+\alias{get_gene_locations}
+\title{Get a table of gene locations}
+\usage{
+get_gene_locations(genome = "hg19")
+}
+\arguments{
+\item{genome}{Which genome build to use. Currently, only "hg19" is available.}
+}
+\description{
+Get a table of gene locations
+}
