+++ man/plot_BISCUT_results.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/plot_BISCUT_results.R
+\name{plot_BISCUT_results}
+\alias{plot_BISCUT_results}
+\title{Visualize BISCUT peaks}
+\usage{
+plot_BISCUT_results(
+  plot_data,
+  title,
+  xlab,
+  abslocs = get_chromosome_coordinates()
+)
+}
+\arguments{
+\item{plot_data}{A data.table (or data.frame) with peak plotting information. For format, see the
+text files in the \code{summary_plots} subdirectory of a BISCUT output directory.}
+}
+\value{
+A ggplot, or NULL if the plot_data is empty.
+}
+\description{
+Summary plots are generated automatically with a BISCUT run. If you like, you can use this function
+to generate custom plots for arbitrary sets of peaks.
+}
