+++ man/plot_fig2.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/plot_fig2.R
+\name{plot_fig2}
+\alias{plot_fig2}
+\title{Plots figure 2}
+\usage{
+plot_fig2(results_dir, abslocs = get_chromosome_coordinates())
+}
+\description{
+Plots figure 2
+}
