+++ tests/testthat/test-first.R
@@ -1,0 +1,7 @@
+
+# testthat placeholder test
+test_that("multiplication works", {
+  expect_equal(2 * 2, 4)
+})
+
+# To-do: Add tests once current outputs are deemed acceptable
